project/python/requirements.txt
websockets
mediapipe
opencv-python
numpy
google-generativeai
openai
python-dotenv
Pillow
requests
